requirements.txt
psycopg2-binary         
SQLAlchemy             
alembic
pydantic                
python-dotenv
loguru